requirements.txt
click
Sphinx
coverage
awscli
flake8
flask
pytest
pytest-isort
pytest-flask
pytest-cov
pytest-flake8
python-dotenv>=0.5.1
